+++ Controllers/UselessController.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using VidsNet.DataModels;
 using VidsNet.Models;
 
 namespace VidsNet.Controllers
@@ -11,8 +12,8 @@ namespace VidsNet.Controllers
     {
         private ILogger _logger;
         private DatabaseContext _db;
-        public SettingsController(ILoggerFactory logger, DatabaseContext db, IHttpContextAccessor accessor) 
-        : base(accessor, db) {
+        public SettingsController(ILoggerFactory logger, DatabaseContext db, UserData userData) 
+        : base(userData) {
             _logger = logger.CreateLogger("Itemsontroller");
             _db = db;
         }