+++ Enums/SubtitleType.cs
@@ -5,6 +5,7 @@ namespace VidsNet.Enums
 {
     public class SubtitleType {
         private List<string> _types {get;set;}
+        private Dictionary<string, string> _mimes {get;set;}
 
         public SubtitleType() {
             _types = new List<string>();
@@ -19,10 +20,32 @@ namespace VidsNet.Enums
             _types.Add(".psb");
             _types.Add(".sami");
             _types.Add(".pjs");
+            _mimes = new Dictionary<string, string>();
+            _mimes.Add(".srt", "application/x-subrip");
+            _mimes.Add(".ssa", "application/octet-stream");
+            _mimes.Add(".ass", "application/octet-stream");
+            _mimes.Add(".smi", "application/smil+xml");
+            _mimes.Add(".sub", "text/vnd.dvb.subtitle");
+            _mimes.Add(".idx", "application/octet-stream");
+            _mimes.Add(".mpl", "application/octet-stream");
+            _mimes.Add(".vtt", "text/vtt");
+            _mimes.Add(".psb", "application/vnd.3gpp.pic-bw-small");
+            _mimes.Add(".sami", "application/octet-stream");
+            _mimes.Add(".pjs", "application/octet-stream");
         }
 
         public bool IsSubtitle(string extension) {
             return _types.Contains(extension);
+        }
+
+        public string GetMime(string extension) {
+            if(_mimes.ContainsKey(extension)) {
+                return _mimes[extension];
+            }
+            else {
+                return "application/octet-stream";
+            }
+            
         }
     }
 }